pyproject.toml
[tool.poetry]
name = "api-for-posts-and-coments"
version = "0.1.0"
description = ""
authors = ["Olesia0402 <olesyashevchuk0402@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
redis = "^5.0.7"
fastapi-limiter = "^0.1.6"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
libgravatar = "^1.0.4"
cloudinary = "^1.40.0"
passlib = "^1.7.4"
fastapi-mail = "^1.4.1"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
setuptools = "^72.1.0"
scikit-learn = "^1.5.1"
joblib = "^1.4.2"
pycryptodome = "^3.20.0"
python-jose = {extras = ["cryptodome"], version = "^3.3.0"}
celery = "^5.4.0"
alembic = "^1.13.2"
bcrypt = "^4.2.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
aiosqlite = "^0.20.0"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
